Finish runLink Matches (#36)
* Match runlinkGetAddressInfo

* Match runlinkTick

* Match runlinkResumeAll

* Match runlinkEnsureJumpIsValid

* MAtch runlinkResumeCode

* Fix kiosk

* Remove if 1

* formatter

* Migrate bss

* Migrate data

* Do some clenaup of the header file

* Make refractOutputAssembler hasm

* Make refractOutputAssembler a proper hasm file with migrated bss

* MAtch runlinkFreeCode

* cleanup work

* tidy tidy

* Name the rest of the data vars for runLink

* Match ProcessRelocationEntry

* Fix PatchInstruction

* Match runlinkInitialise

* Match runLink

* Fix romCopy and piRomLoadSection address to be ptrs.

* Make vrambase a void pointer.

* pending overlays now have vrambase void as well

* Fix to using u32 for overlay sizes

* Make an attempt to migrate some rodata

* Formatter

+++ src/level.c
@@ -84,7 +84,7 @@ void levelGetCounts(void) {
 
     D_800FB118_B5958 = hdrBuf;
     for (i = 0; i < D_800FB124_B1764; i++) {
-        piRomLoadSection(0x1F, (u32) D_800FB118_B5958, D_800FB110_B1750[i], sizeof(LevelHeader) + 4);
+        piRomLoadSection(0x1F, D_800FB118_B5958, D_800FB110_B1750[i], sizeof(LevelHeader) + 4);
 
         if (D_800FB128_B1768 < D_800FB118_B5958->unk20) {
             D_800FB128_B1768 = D_800FB118_B5958->unk20;
